Add namespace metadata option in templates
This sets the namespace key in metadata for all kubernetes manifests.
The `.Release.Namespace` is a built-in variable in helm, that can be
passed with` --namespace` on `install` or `template` step. If not set
it defaults to `default`.

*Why?*

When using `helm install` with the `--namespace` argument the templates
are applied into the chosen namespace.

When using helm only as a template engine, in combination with GitOps
tools like Flux, the namespace property must be set in the rendered
manifests to be applied. For `--namespace` to have an effect when using
`helm template`, it must be present like this in the templates.

This change should not have any effect/change for `helm install`
approach as far as I know.